refactor(Restaurants): 파생 상태를 만드는 메서드 filter, sort를 RestaurantList 컴포넌트로 이동
도메인에서 해당 메서드 분리
